v2 accomplish

+++ StudentInformationSystem-be/src/model/Graduate.java
@@ -5,7 +5,7 @@ public class Graduate extends Student {
     private String researchArea; // 研究方向
 
 
-    public Graduate(int id, String name, int age, Address address, String classNum, String advisor, String researchArea, double mathScore, double englishScore, double historyScore) {
+    public Graduate(long id, String name, int age, Address address, String classNum, String advisor, String researchArea, double mathScore, double englishScore, double historyScore) {
         super(id, name, age, address, classNum);
         scores.add(new Score("Math", mathScore));
         scores.add(new Score("English", englishScore));
